workaround: allow any breadcrumb level to be clickable
This is a workaround because our spaces/mountpoint implementation is not correct. But since we expose the root of a user as a space, which then makes the share have the full internal path, this means there are parts of the path the user doesn't have access to.
But for now is ok to let them try to access to paths and get a denied page.